Fix errors from gatsby inference and populate=deep
populate=deep is incompatible gatsby-source-strapi. Also,
gatsby-source-strapi is stupid.

- When loading Strapi Components of "regular" objects,
  gatsby-source-strapi converts them to Gatsby Nodes recursively using
  createNode.
- This also renames e.g. the `content` field of `Page` to
  `content___NODE` to mark it as a Gatsby relation.
- When loading relation fields, this doesn't happen. Instead,
  gatsby-source-strapi just createNode()s the related object as-is,
  but only if it doesn't exist.
- If `LandingPage` is populated deeply, the `buttonLink` relation
  contains a `Page`, which further has the `content` field populated.
- If `LandingPage` then ends up loaded before `Page`, that related
  `Page` is createNode()'d with its `content` as JSON (not converted to
  Nodes).
- After this, when `Page`s are loaded, they're created normally as
  STRAPI_PAGE Nodes with a `content___NODE` field, which would also
  appear in GraphQL as `content`. (This shows up as a console warning.)
- This `content` mismatch messes up Gatsby's node creation or inference
  or something, and the `buttonLink`ed STRAPI_PAGE ends up with null
  content. I spent a few hours tracing through this spaghetti code.

This commit explicitly specifies the population to avoid the above
clusterfuck. It still populates `buttonLink`, but only with `id`.

This *seems* to work fine for now. I'm not sure of the implications of
the junk Node created by the relation: I'm not sure if it's overwritten
or merged with the real STRAPI_PAGE created later. This has implications
later for getting `buttonLink` to work at all in a reasonable way.

In general, Strapi relations (especially as handled by
gatsby-source-strapi) don't seem to work well with Gatsby's GraphQL.